backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
supabase
yfinance
xgboost
shap
transformers
torch
requests
praw
scikit-learn
pandas
numpy
textblob
vaderSentiment
